ReadFromRelationalDB transform re-implemeted using a ParDo
