+++ beam_nuggets/io/__init__.py
@@ -1,7 +1,5 @@
 from __future__ import division, print_function
 
 from csv_ import ReadFromCsv
-from relational_db.relational_db_transform import (
-    WriteToRelationalDB,
-    ReadFromRelationalDB
-)
+from relational_db.read_transform import ReadFromRelationalDB
+from relational_db.write_transform import WriteToRelationalDB
